==> 03_tweets_to_pickle.ipynb <==
__Subject:__ Topics of Tweets of US Congress through 
             the Lens of Campaign Finance

__Date:__ 09/08/2018

__Name:__ Edmund D. Chitwood

***

__Summary:__<br>
<br>The following Notebook 
- reads all Tweets from CSV files,
- reads them into a Pandas DataFrame,and
- then pickles the DataFrame.
***

import pandas as pd
import pickle

# Navigate to data folder which contains csv files
# with Tweets and Tweet metadata by all members of 
# Congress and get names of all csv files
csv_names = !var="cd data && ls" && eval $var
csv_names = list(csv_names)

# Initialize DataFrame to hold Tweets and Tweet metadata
tweets = pd.read_csv('data/'+csv_names[0],sep=';',header=None)

# Read in data for all members of Congress
for i in range(1,len(csv_names)):
    new_tweets = pd.read_csv('data/'+csv_names[i],sep=';',header=None)
    tweets = tweets.append(new_tweets)

tweets.columns = (['username','date','retweets','favorites','text','mentions',\
                   'hashtags','id','permalink'])

tweets.head()

tweets.to_pickle('congress_tweets.pkl')
