==> 05_congress_metadata.ipynb <==
__Subject:__ Topics of Tweets of US Congress through the Lens of Campaign Finance

__Date:__ 09/12/2018

__Name:__ Edmund D. Chitwood
***

__Summary:__<br>
<br>The following Notebook
- downloads metadata for each Congress member,
- resolves discrepancies between the Twitter usernames listed in the Congress metadata and Tweet data.
***

import pandas as pd
import pickle

# Download CSV file that contains geographic, party, social media, etc. data
# about members of the US Congress. Source:
# https://theunitedstates.io/congress-legislators/legislators-current.csv

congress_metadata = pd.read_csv('legislators-current.csv')
congress_metadata.columns

# Indicate Congress members for whom no Twitter username was listed
congress_metadata.twitter.fillna(value='no twitter',inplace=True)

congress_metadata[congress_metadata.twitter=='no twitter']

# Lowercase Twitter usernames in order to join with Tweet data
congress_metadata.twitter = congress_metadata.twitter.apply(lambda x:x.lower()) 

# Import Tweets to compare usernames to Congress metadata Twitter usernames
tweets = pd.read_pickle('preprocessed_tweets.pkl')

tweets.head()

# Lowercase Twitter usernames in order to join with Congress metadata
tweets.username = tweets.username.apply(lambda x:x.lower()) 

# Compare Twitter usernames in Congress metadata to those in Tweets DataFrame
cm_usernames = set(congress_metadata.twitter)
tweets_usernames = set(tweets.username.unique())

in_cm_not_tweets = cm_usernames - tweets_usernames
in_tweets_not_cm = tweets_usernames - cm_usernames

in_cm_not_tweets

in_tweets_not_cm

# The following Twitter usernames were listed in Twitter username source: 
# https://gwu-libraries.github.io/sfm-ui/posts/2017-05-23-congress-seed-list
# but their owners were no longer in Congress as of 09/12/18
former_congress_members= ['farenthold','jasoninthehouse','louiseslaughter',\
                          'lutherstrange','patmeehanpa', 'pattiberi', \
                          'repbecerra', 'repcharliedent','repjbridenstine',\
                          'repjohnconyers','senfranken', 'senjohnmccain', 'senthadcochran']

tweets.username.nunique()

# Drop Tweets from preprocessed Tweets if Tweeters are no longer in Congress
for i in former_congress_members:
    tweets.reset_index(inplace=True,drop=True)
    tweets.drop(tweets[tweets.username==i].index,inplace=True)
    

tweets.to_pickle('updated_preprocessed_tweets.pkl')

# There were discrepancies between the two data sources for the following usernames:

twitter_usernames_tweets = (['maxinewaters','kamalaharris','gregorymeeks',\
                            'devinnunes','davidrouzer','danarohrabacher',\
                            'brettguthrie','anthonybrownmd4',\
                            'scotttaylorva','rogermarshallmd',\
                            'repblumenauer','amyklobuchar'])

twitter_usernames_cm = (['repmaxinewaters','senkamalaharris',\
                         'repgregorymeeks','repdevinnunes',\
                         'repdavidrouzer','reprohrabacher','repguthrie',\
                         'repanthonybrown','repscotttaylor','repmarshall',\
                         'blumenauermedia','senamyklobuchar'])


# Change names in T.U.C. to the corresponding ones in T.U.T. for purposes of 
# joining Tweet data to Congress metadata

for i in range(len(twitter_usernames_cm)):
    index = (congress_metadata[congress_metadata.twitter==\
                    twitter_usernames_cm\
                    [i]].twitter.index)
    congress_metadata.at[index, 'twitter'] = twitter_usernames_tweets[i]


# The Congress metadata source was missing Twitter usernames for
# several current members of Congress who have Twitter accounts
# Add Twitter names to Congress members with nan's where applicable

def add_username(last_name,username):
    
    index = (congress_metadata[congress_metadata.last_name==\
                    last_name]).index

    congress_metadata.at[index, 'twitter'] = username

add_username('Cassidy','billcassidy')
add_username('Amash','justinamash')
add_username('Comer','kycomer')
add_username('Kaine','timkaine')

# Compare Twitter usernames in Congress metadata to those in Tweets DataFrame
cm_usernames = set(congress_metadata.twitter)
tweets_usernames = set(tweets.username.unique())

in_cm_not_tweets = cm_usernames - tweets_usernames
in_tweets_not_cm = tweets_usernames - cm_usernames

# Discrepancies in Tweets DataFrame resolved
in_tweets_not_cm

# Unresolved Discrepancies in Congress metadata
in_cm_not_tweets

# Drop the Congress members not accounted for in Tweets DataFrame from Congress 
# metadata
for i in in_cm_not_tweets:
    congress_metadata.reset_index(inplace=True,drop=True)
    (congress_metadata.drop(congress_metadata[congress_metadata\
                        .twitter==i].index,inplace=True))
    

tweets.username.nunique()

# Twitter usernames in two DataFrames now match 
congress_metadata.twitter.nunique()

congress_metadata.to_pickle('congress_metadata.pkl')
