==> 07_campaign_finance_data.ipynb <==
__Subject:__ Topics of Tweets of US Congress through the Lens of Campaign Finance

__Date:__ 09/12/2018

__Name:__ Edmund D. Chitwood
***

__Summary:__<br>
<br>The following Notebook
- downloads campaign finance information for each relevant member of Congress from opensecrets.org
***

import pandas as pd
import pickle

congress_metadata = pd.read_pickle('congress_metadata.pkl')

congress_metadata.head()

congress_metadata.columns

# https://www.opensecrets.org/ is data source for campaign finance data
#!pip install opensecrets-crpapi

from crpapi import CRP
#crp = CRP(apikey)
crp = CRP('21539018e701b26436921b72553a4208')

# Example campaign finance data
contributions = crp.candidates.sector('N00003535')
contributions

# Get Campaign contriubtion information for each member of Congress, broken 
# down by sector and type of donor (i.e. Individual or PAC) 
# Totals are from most recent election cycle

contributions = []

for i in congress_metadata.opensecrets_id:
    try:
        contributions.append(crp.candidates.sector(i))
    except:
        print(i)
        contributions.append('empty')

# No data returned for Representative Chris Collins:
# https://www.opensecrets.org/members-of-congress/summary?cid=N00001285&cycle=2018
# Collins suspended his campaign on August 11, 2018 after he was indicted for
# insider trading 
# Drop him from data

contributions.remove('empty')
congress_metadata.drop(congress_metadata[congress_metadata.opensecrets_id\
                       =='N00000267'].index, inplace=True)

# JSON objects for each member of Congress vary
# For most Congress members, JSON object contain info for 13 sectors
# That is not the case for the following
for v in contributions:
    if len(v) != 13:
        print(len(v))

# Flatten list of dicts of dicts
flat_contributions = []

for i,v in enumerate(contributions):
    if type(v) == list:
        for j in range(len(v)):
            contributions[i][j]['@attributes']['os_id'] = ids[i]
            flat_contributions.append(contributions[i][j]['@attributes'])
    else:
        v['@attributes']['os_id'] = ids[i]
        flat_contributions.append(v['@attributes'])

flat_contributions[0]

contributions_congress = pd.DataFrame(flat_contributions)

contributions_congress.shape

contributions_congress.describe()

contributions_congress.to_pickle('contributions_congress.pkl')
