==> 08_campaign_finance_eda_aggregate_pac.ipynb <==
__Subject:__ Topics of Tweets of US Congress through the Lens of Campaign Finance

__Date:__ 09/13/2018

__Name:__ Edmund D. Chitwood
***

__Summary:__<br>
<br>The following Notebook
- perform EDA on aggreate camapaign contribution data for current Congress, and
- aggregate PAC contribution totals for each relevant member of Congress
***

import pandas as pd
import pickle

contributions = pd.read_pickle('contributions_congress.pkl')

contributions.shape

contributions.total = contributions.total.apply(lambda x: int(x))
contributions.indivs = contributions.indivs.apply(lambda x: int(x))
contributions.pacs = contributions.pacs.apply(lambda x: int(x))

contributions.describe()

contributions.total.sum()

contributions.sector_name.unique()

contributions[contributions.sector_name=='Construction'].shape

pac_money_sector_totals = dict()

for i in contributions.sector_name.unique():
    pac_money_sector_totals[i] = contributions[contributions.sector_name==i].pacs.sum()

sorted(pac_money_sector_totals.items(), key=lambda kv: kv[1],reverse=True)

money_sector_totals = dict()

for i in contributions.sector_name.unique():
    money_sector_totals[i] = contributions[contributions.sector_name==i].total.sum()

sorted(money_sector_totals.items(), key=lambda kv: kv[1],reverse=True)

sector_by_number_congressmen = dict()

for i in contributions.sector_name.unique():
    sector_by_number_congressmen[i] = contributions[contributions.sector_name==i].shape

sorted(sector_by_number_congressmen.items(), key=lambda kv: kv[1],reverse=True)

contributions[contributions['os_id']=='N00003535'].pacs.sum()

contributions.shape

grouped = contributions.groupby('os_id')

os_id = []
pacs = []

for name, group in grouped:
    os_id.append(name)
    pacs.append(group[group['os_id']==name].pacs.sum())

contributions_totals = pd.DataFrame()

contributions_totals['os_id'] = os_id
contributions_totals['pacs'] = pacs

contributions_totals.head()

contributions_totals.describe()

contributions_totals.pacs.median()

contributions_totals.pacs.mean()

contributions_totals.to_pickle('contributions_pacs.pkl')
